refactor!: Remove 'Format' namespace and use named exports
`--experimental-strip-types` 対応として。
